run each AMI build in its own ods-core clone

+++ ods-devenv/buildbot/scripts/runAmiBuild.sh
@@ -6,6 +6,8 @@
 set -exu
 
 targetGitRef="master"
+readonly repository=ods-core
+readonly odsCoreClonePath=https://github.com/opendevstack/ods-core
 
 while [[ "$#" -gt 0 ]]; do
   case $1 in
@@ -22,13 +24,37 @@ esac; shift; done
 export PACKER_LOG=1
 export AWS_MAX_ATTEMPTS=400
 export AWS_POLL_DELAY_SECONDS=15
-# source "${HOME}/.packerrc"
-
-cd "${buildPath:?}" || exit
 date
+
+mkdir -p "${buildPath:?}"
+cd "${buildPath}" || exit
+
+# prepare build folder
+readonly targetGitRefForPath=$(echo -n "${targetGitRef:?}" | tr "/" "_")
+if [[ -d "${targetGitRefForPath}" ]]
+then
+  # clean up stale folder from previous builds
+  rm -rf "${targetGitRefForPath}"
+fi
+mkdir -p "${targetGitRefForPath}"
+cd "${targetGitRefForPath}"
+git clone "${odsCoreClonePath}"
+cd "${repository}"
+if ! git checkout "${targetGitRef}" &> /dev/null
+then
+  echo "Branch to build ${targetGitRef} could not be checked out, please fix the configuration!"
+  exit 1
+fi
+
+# prepare logs
 mkdir -p "${logPath:?}"
-logFile="${logPath:?}/build_$(echo "${targetGitRef}" | tr "/" "_")_$(date +%Y%m%dT%H%M%S).log"
-targetGitRefForPath=$(echo -n "${targetGitRef}" | tr "/" "_")
+readonly logFile="${logPath:?}/build_$(echo "${targetGitRef:?}" | tr "/" "_")_$(date +%Y%m%dT%H%M%S).log"
 rm -f "${logPath}/current_${targetGitRefForPath}.log"
 ln -s "${logFile}" "${logPath}/current_${targetGitRefForPath}.log"
-time bash 2>&1 ods-devenv/packer/create_ods_box_image.sh --target create_ods_box_ami --aws-access-key "${awsAccessKey}" --aws-secret-key "${awsSecretAccessKey}" --ods-branch "${targetGitRef}" --instance-type "${instanceType}" | tee "${logFile:?}"
+
+# run packer build
+time bash 2>&1 ods-devenv/packer/create_ods_box_image.sh --target create_ods_box_ami --aws-access-key "${awsAccessKey:?}" --aws-secret-key "${awsSecretAccessKey:?}" --ods-branch "${targetGitRef}" --instance-type "${instanceType:?}" | tee "${logFile}"
+
+# clean up after build
+cd ..
+rm -rf "${targetGitRefForPath}"
